#  optimize some hotspot use cases of deltaR and polar in reco/miniAOD (backport of #29918 )