--- a/PhysicsTools/PatAlgos/plugins/PATElectronProducer.cc
+++ b/PhysicsTools/PatAlgos/plugins/PATElectronProducer.cc
@@ -973,12 +973,12 @@ void PATElectronProducer::setElectronMiniIso(Electron& anElectron, const PackedC
 {
   pat::PFIsolation miniiso;
   if(anElectron.isEE())
-      miniiso = pat::getMiniPFIsolation(pc, anElectron.p4(),
+      miniiso = pat::getMiniPFIsolation(pc, anElectron.polarP4(),
                                         miniIsoParamsE_[0], miniIsoParamsE_[1], miniIsoParamsE_[2],
                                         miniIsoParamsE_[3], miniIsoParamsE_[4], miniIsoParamsE_[5],
                                         miniIsoParamsE_[6], miniIsoParamsE_[7], miniIsoParamsE_[8]);
   else
-      miniiso = pat::getMiniPFIsolation(pc, anElectron.p4(),
+      miniiso = pat::getMiniPFIsolation(pc, anElectron.polarP4(),
                                         miniIsoParamsB_[0], miniIsoParamsB_[1], miniIsoParamsB_[2],
                                         miniIsoParamsB_[3], miniIsoParamsB_[4], miniIsoParamsB_[5],
                                         miniIsoParamsB_[6], miniIsoParamsB_[7], miniIsoParamsB_[8]);