--- a/PhysicsTools/PatAlgos/plugins/PATIsolatedTrackProducer.cc
+++ b/PhysicsTools/PatAlgos/plugins/PATIsolatedTrackProducer.cc
@@ -43,6 +43,7 @@ namespace pat {
 
     class PATIsolatedTrackProducer : public edm::stream::EDProducer<> {
       public:
+          typedef pat::IsolatedTrack::PolarLorentzVector PolarLorentzVector;
           typedef pat::IsolatedTrack::LorentzVector LorentzVector;
 
           explicit PATIsolatedTrackProducer(const edm::ParameterSet&);
@@ -51,20 +52,20 @@ namespace pat {
           void produce(edm::Event&, const edm::EventSetup&) override;
         
           // compute iso/miniiso
-          void getIsolation(const LorentzVector& p4, const pat::PackedCandidateCollection* pc, int pc_idx,
+          void getIsolation(const PolarLorentzVector& p4, const pat::PackedCandidateCollection* pc, int pc_idx,
                             pat::PFIsolation &iso, pat::PFIsolation &miniiso) const;
 
-          bool getPFLeptonOverlap(const LorentzVector& p4, const pat::PackedCandidateCollection* pc) const;
+          bool getPFLeptonOverlap(const PolarLorentzVector& p4, const pat::PackedCandidateCollection* pc) const;
 
-          float getPFNeutralSum(const LorentzVector& p4, const pat::PackedCandidateCollection* pc, int pc_idx) const;
+          float getPFNeutralSum(const PolarLorentzVector& p4, const pat::PackedCandidateCollection* pc, int pc_idx) const;
 
-          void getNearestPCRef(const LorentzVector& p4, const pat::PackedCandidateCollection *pc, int pc_idx, int& pc_ref_idx) const;
+          void getNearestPCRef(const PolarLorentzVector& p4, const pat::PackedCandidateCollection *pc, int pc_idx, int& pc_ref_idx) const;
       
           float getDeDx(const reco::DeDxHitInfo *hitInfo, bool doPixel, bool doStrip) const ;
 
           TrackDetMatchInfo getTrackDetMatchInfo(const edm::Event&, const edm::EventSetup&, const reco::Track&);
 
-          void getCaloJetEnergy(const LorentzVector&, const reco::CaloJetCollection*, float&, float&) const;
+          void getCaloJetEnergy(const PolarLorentzVector&, const reco::CaloJetCollection*, float&, float&) const;
 
       private:             
           const edm::EDGetTokenT<pat::PackedCandidateCollection>    pc_;
@@ -239,6 +240,7 @@ void pat::PATIsolatedTrackProducer::produce(edm::Event& iEvent, const edm::Event
         bool isInPackedCands = (pcref.isNonnull() && pcref.id()==pc_h.id() && pfCand.charge()!=0);
         bool isInLostTracks  = (ltref.isNonnull() && ltref.id()==lt_h.id());
 
+        PolarLorentzVector polarP4;
         LorentzVector p4;
         pat::PackedCandidateRef refToCand;
         int pdgId, charge, fromPV;
@@ -249,11 +251,13 @@ void pat::PATIsolatedTrackProducer::produce(edm::Event& iEvent, const edm::Event
         // get the four-momentum and charge
         if(isInPackedCands){
             p4        = pfCand.p4();
+            polarP4        = pfCand.polarP4();
             charge    = pfCand.charge();
             pfCandInd = pcref.key();
 	    ltCandInd = -1;
         }else if(isInLostTracks){
             p4        = lostTrack.p4();
+            polarP4        = lostTrack.polarP4();
             charge    = lostTrack.charge();
             pfCandInd = -1;
 	    ltCandInd = ltref.key();
@@ -261,6 +265,7 @@ void pat::PATIsolatedTrackProducer::produce(edm::Event& iEvent, const edm::Event
             double m = 0.13957018; //assume pion mass
             double E = sqrt(m*m + gentk.p()*gentk.p());
             p4.SetPxPyPzE(gentk.px(), gentk.py(), gentk.pz(), E);
+            polarP4.SetCoordinates(gentk.pt(), gentk.eta(), gentk.phi(), m);
             charge = gentk.charge();
             pfCandInd = -1;
 	    ltCandInd = -1;
@@ -274,7 +279,7 @@ void pat::PATIsolatedTrackProducer::produce(edm::Event& iEvent, const edm::Event
             }
         }
 
-        if(p4.pt() < pT_cut_ && prescaled <= 1) 
+        if(polarP4.pt() < pT_cut_ && prescaled <= 1) 
             continue;
         if(charge == 0)
             continue;
@@ -282,13 +287,13 @@ void pat::PATIsolatedTrackProducer::produce(edm::Event& iEvent, const edm::Event
         // get the isolation of the track
         pat::PFIsolation isolationDR03;
         pat::PFIsolation miniIso;
-        getIsolation(p4, pc, pfCandInd, isolationDR03, miniIso);
+        getIsolation(polarP4, pc, pfCandInd, isolationDR03, miniIso);
         
         // isolation cut
-        if( p4.pt() < pT_cut_noIso_ && prescaled <= 1 && 
+        if( polarP4.pt() < pT_cut_noIso_ && prescaled <= 1 && 
             !(isolationDR03.chargedHadronIso() < absIso_cut_ ||
-              isolationDR03.chargedHadronIso()/p4.pt() < relIso_cut_ ||
-              miniIso.chargedHadronIso()/p4.pt() < miniRelIso_cut_))
+              isolationDR03.chargedHadronIso()/polarP4.pt() < relIso_cut_ ||
+              miniIso.chargedHadronIso()/polarP4.pt() < miniRelIso_cut_))
             continue;
 
         // get the rest after the pt/iso cuts. Saves some runtime
@@ -319,20 +324,20 @@ void pat::PATIsolatedTrackProducer::produce(edm::Event& iEvent, const edm::Event
         }
 
         float caloJetEm, caloJetHad;
-        getCaloJetEnergy(p4, caloJets.product(), caloJetEm, caloJetHad);
+        getCaloJetEnergy(polarP4, caloJets.product(), caloJetEm, caloJetHad);
 
-	bool pfLepOverlap = getPFLeptonOverlap(p4, pc);
-	float pfNeutralSum = getPFNeutralSum(p4, pc, pfCandInd);
+	bool pfLepOverlap = getPFLeptonOverlap(polarP4, pc);
+	float pfNeutralSum = getPFNeutralSum(polarP4, pc, pfCandInd);
 	
 	pat::PackedCandidateRef refToNearestPF = pat::PackedCandidateRef();
 	int refToNearestPF_idx=-1;
-	getNearestPCRef(p4, pc, pfCandInd, refToNearestPF_idx);
+	getNearestPCRef(polarP4, pc, pfCandInd, refToNearestPF_idx);
 	if(refToNearestPF_idx!=-1)
 	  refToNearestPF = pat::PackedCandidateRef(pc_h, refToNearestPF_idx);
 	
 	pat::PackedCandidateRef refToNearestLostTrack = pat::PackedCandidateRef();
 	int refToNearestLostTrack_idx=-1;
-	getNearestPCRef(p4, lt, ltCandInd, refToNearestLostTrack_idx);
+	getNearestPCRef(polarP4, lt, ltCandInd, refToNearestLostTrack_idx);
 	if(refToNearestLostTrack_idx!=-1)
 	  refToNearestLostTrack = pat::PackedCandidateRef(lt_h, refToNearestLostTrack_idx);
 	
@@ -405,15 +410,15 @@ void pat::PATIsolatedTrackProducer::produce(edm::Event& iEvent, const edm::Event
         if(pfref.get()->trackRef().isNonnull() && pfref.get()->trackRef().id() == gt_h.id())
             continue;
 
-        LorentzVector p4;
+        PolarLorentzVector polarP4;
         pat::PackedCandidateRef refToCand;
         int pdgId, charge, fromPV;
         float dz, dxy, dzError, dxyError;
 
-        p4 = pfCand.p4();
+        polarP4 = pfCand.polarP4();
         charge = pfCand.charge();
 
-        if(p4.pt() < pT_cut_)
+        if(polarP4.pt() < pT_cut_)
             continue;
         if(charge == 0)
             continue;
@@ -421,13 +426,13 @@ void pat::PATIsolatedTrackProducer::produce(edm::Event& iEvent, const edm::Event
         // get the isolation of the track
         pat::PFIsolation isolationDR03;
         pat::PFIsolation miniIso;
-        getIsolation(p4, pc, ipc, isolationDR03, miniIso);
+        getIsolation(polarP4, pc, ipc, isolationDR03, miniIso);
         
         // isolation cut
-        if( p4.pt() < pT_cut_noIso_ && 
+        if( polarP4.pt() < pT_cut_noIso_ && 
             !(isolationDR03.chargedHadronIso() < absIso_cut_ ||
-              isolationDR03.chargedHadronIso()/p4.pt() < relIso_cut_ ||
-              miniIso.chargedHadronIso()/p4.pt() < miniRelIso_cut_))
+              isolationDR03.chargedHadronIso()/polarP4.pt() < relIso_cut_ ||
+              miniIso.chargedHadronIso()/polarP4.pt() < miniRelIso_cut_))
             continue;
 
         pdgId     = pfCand.pdgId();
@@ -444,20 +449,20 @@ void pat::PATIsolatedTrackProducer::produce(edm::Event& iEvent, const edm::Event
         refToCand = pcref;
 
         float caloJetEm, caloJetHad;
-        getCaloJetEnergy(p4, caloJets.product(), caloJetEm, caloJetHad);
+        getCaloJetEnergy(polarP4, caloJets.product(), caloJetEm, caloJetHad);
 
-	bool pfLepOverlap = getPFLeptonOverlap(p4, pc);
-	float pfNeutralSum = getPFNeutralSum(p4, pc, ipc);
+	bool pfLepOverlap = getPFLeptonOverlap(polarP4, pc);
+	float pfNeutralSum = getPFNeutralSum(polarP4, pc, ipc);
 
 	pat::PackedCandidateRef refToNearestPF = pat::PackedCandidateRef();
 	int refToNearestPF_idx=-1;
-	getNearestPCRef(p4, pc, ipc, refToNearestPF_idx);
+	getNearestPCRef(polarP4, pc, ipc, refToNearestPF_idx);
 	if(refToNearestPF_idx!=-1)
 	  refToNearestPF = pat::PackedCandidateRef(pc_h, refToNearestPF_idx);
 	
 	pat::PackedCandidateRef refToNearestLostTrack = pat::PackedCandidateRef();
 	int refToNearestLostTrack_idx=-1;
-	getNearestPCRef(p4, lt, -1, refToNearestLostTrack_idx);
+	getNearestPCRef(polarP4, lt, -1, refToNearestLostTrack_idx);
 	if(refToNearestLostTrack_idx!=-1)
 	  refToNearestLostTrack = pat::PackedCandidateRef(lt_h, refToNearestLostTrack_idx);
 
@@ -470,7 +475,7 @@ void pat::PATIsolatedTrackProducer::produce(edm::Event& iEvent, const edm::Event
         int deltaEta=0;
         int deltaPhi=0;
 
-        outPtrP->push_back(pat::IsolatedTrack(isolationDR03, miniIso, caloJetEm, caloJetHad, pfLepOverlap, pfNeutralSum, p4,
+        outPtrP->push_back(pat::IsolatedTrack(isolationDR03, miniIso, caloJetEm, caloJetHad, pfLepOverlap, pfNeutralSum, pfCand.p4(),
                                               charge, pdgId, dz, dxy, dzError, dxyError,
                                               hp, dEdxStrip, dEdxPixel, fromPV, trackQuality,
                                               ecalStatus, hcalStatus, deltaEta, deltaPhi, refToCand,
@@ -492,7 +497,7 @@ void pat::PATIsolatedTrackProducer::produce(edm::Event& iEvent, const edm::Event
 }
 
 
-void pat::PATIsolatedTrackProducer::getIsolation(const LorentzVector& p4, 
+void pat::PATIsolatedTrackProducer::getIsolation(const PolarLorentzVector& p4, 
                                                  const pat::PackedCandidateCollection *pc, int pc_idx,
                                                  pat::PFIsolation &iso, pat::PFIsolation &miniiso) const
 {
@@ -505,7 +510,7 @@ void pat::PATIsolatedTrackProducer::getIsolation(const LorentzVector& p4,
             int id = std::abs(pf_it->pdgId());
             bool fromPV = (pf_it->fromPV()>1 || fabs(pf_it->dz()) < pfIsolation_DZ_);
             float pt = pf_it->p4().pt();
-            float dr = deltaR(p4, pf_it->p4());
+            float dr = deltaR(p4, *pf_it);
 
             if(dr < pfIsolation_DR_){
                 // charged cands from PV get added to trackIso
@@ -540,7 +545,7 @@ void pat::PATIsolatedTrackProducer::getIsolation(const LorentzVector& p4,
 }
 
 //get overlap of isolated track with a PF lepton
-bool pat::PATIsolatedTrackProducer::getPFLeptonOverlap(const LorentzVector& p4, const pat::PackedCandidateCollection *pc) const
+bool pat::PATIsolatedTrackProducer::getPFLeptonOverlap(const PolarLorentzVector& p4, const pat::PackedCandidateCollection *pc) const
 {
         bool isOverlap = false;
 	float dr_min = pflepoverlap_DR_;
@@ -557,7 +562,7 @@ bool pat::PATIsolatedTrackProducer::getPFLeptonOverlap(const LorentzVector& p4, 
 	    if(pt < pflepoverlap_pTmin_) // exclude pf candidates w/ pT below threshold
 	      continue;
 
-            float dr = deltaR(p4, pf.p4());
+            float dr = deltaR(p4, pf);
             if(dr > pflepoverlap_DR_) // exclude pf candidates far from isolated track
 	      continue;
 
@@ -575,7 +580,7 @@ bool pat::PATIsolatedTrackProducer::getPFLeptonOverlap(const LorentzVector& p4, 
 }
 
 //get ref to nearest pf packed candidate
-void pat::PATIsolatedTrackProducer::getNearestPCRef(const LorentzVector& p4, const pat::PackedCandidateCollection *pc, int pc_idx, int& pc_ref_idx) const
+void pat::PATIsolatedTrackProducer::getNearestPCRef(const PolarLorentzVector& p4, const pat::PackedCandidateCollection *pc, int pc_idx, int& pc_ref_idx) const
 {
 	float dr_min = pcRefNearest_DR_;
 	float dr_min_pu = pcRefNearest_DR_;
@@ -586,7 +591,7 @@ void pat::PATIsolatedTrackProducer::getNearestPCRef(const LorentzVector& p4, con
 	    int charge = std::abs(pf_it->charge());
             bool fromPV = (pf_it->fromPV()>1 || fabs(pf_it->dz()) < pfIsolation_DZ_);
             float pt = pf_it->p4().pt();
-            float dr = deltaR(p4, pf_it->p4());
+            float dr = deltaR(p4, *pf_it);
 	    if(charge==0) // exclude neutral candidates
 	      continue;
 	    if(pt < pcRefNearest_pTmin_) // exclude candidates w/ pT below threshold
@@ -613,7 +618,7 @@ void pat::PATIsolatedTrackProducer::getNearestPCRef(const LorentzVector& p4, con
 }
 
 // get PF neutral pT sum around isolated track
-float pat::PATIsolatedTrackProducer::getPFNeutralSum(const LorentzVector& p4, const pat::PackedCandidateCollection *pc, int pc_idx) const
+float pat::PATIsolatedTrackProducer::getPFNeutralSum(const PolarLorentzVector& p4, const pat::PackedCandidateCollection *pc, int pc_idx) const
 {
         float nsum=0;
         float nhsum=0, phsum=0;
@@ -622,7 +627,7 @@ float pat::PATIsolatedTrackProducer::getPFNeutralSum(const LorentzVector& p4, co
                 continue;
             int id = std::abs(pf_it->pdgId());
             float pt = pf_it->p4().pt();
-            float dr = deltaR(p4, pf_it->p4());
+            float dr = deltaR(p4, *pf_it);
 
             if(dr < pfneutralsum_DR_){
 	      // neutral hadron sum
@@ -698,13 +703,13 @@ TrackDetMatchInfo pat::PATIsolatedTrackProducer::getTrackDetMatchInfo(const edm:
     return trackAssociator_.associate(iEvent, iSetup, trackAssocParameters_, &initialState);
 }
 
-void pat::PATIsolatedTrackProducer::getCaloJetEnergy(const LorentzVector& p4, const reco::CaloJetCollection* cJets,
+void pat::PATIsolatedTrackProducer::getCaloJetEnergy(const PolarLorentzVector& p4, const reco::CaloJetCollection* cJets,
                                                      float &caloJetEm, float& caloJetHad) const
 {
     float nearestDR = 999;
     int ind = -1;
     for(unsigned int i=0; i<cJets->size(); i++){
-        float dR = deltaR(cJets->at(i).p4(), p4);
+        float dR = deltaR(cJets->at(i), p4);
         if(dR < caloJet_DR_ && dR < nearestDR){
             nearestDR = dR;
             ind = i;