--- a/PhysicsTools/PatUtils/interface/MiniIsolation.h
+++ b/PhysicsTools/PatUtils/interface/MiniIsolation.h
@@ -15,18 +15,18 @@
 
 namespace pat{
 
-  float miniIsoDr(const math::XYZTLorentzVector &p4, float mindr, float maxdr,
+  float miniIsoDr(const reco::Candidate::PolarLorentzVector &p4, float mindr, float maxdr,
 		  float kt_scale);
 
   // see src file for definitions of parameters
-  PFIsolation getMiniPFIsolation(const pat::PackedCandidateCollection *pfcands, const math::XYZTLorentzVector& p4,
-                                   float mindr=0.05, float maxdr=0.2, float kt_scale=10.0,
-                                   float ptthresh=0.5, float deadcone_ch=0.0001,
-                                   float deadcone_pu=0.01, float deadcone_ph=0.01, float deadcone_nh=0.01,
-                                   float dZ_cut=0.0);
+  PFIsolation getMiniPFIsolation(const pat::PackedCandidateCollection *pfcands, const reco::Candidate::PolarLorentzVector& p4,
+                                 float mindr=0.05, float maxdr=0.2, float kt_scale=10.0,
+                                 float ptthresh=0.5, float deadcone_ch=0.0001,
+                                 float deadcone_pu=0.01, float deadcone_ph=0.01, float deadcone_nh=0.01,
+                                 float dZ_cut=0.0);
 
   double muonRelMiniIsoPUCorrected(const PFIsolation& iso,
-				   const math::XYZTLorentzVector& p4,
+				   const reco::Candidate::PolarLorentzVector& p4,
 				   double dr,
 				   double rho,
                                    const std::vector<double> &area);