backport of #29918

Technical improvements to reduce time in calls to deltaR where polar Lorentz vectors can be used or where (eta,phi) are available directly.
Here the expected gain is mainly limited to miniIsolation and isolatedTracks. 
In a 2016 re-miniAOD workflow  136.7611 times per event in relevant modules are reduced significantly for a total of about 4% of the re-miniAOD step:
```
2.32 ms/ev ->         0.81 ms/ev patElectrons
115.03 ms/ev ->        51.91 ms/ev isolatedTracks
21.10 ms/ev ->         9.67 ms/ev patMuons

```
no changes are expected based on #29918