@cmsbuild 2020-05-20
A new Pull Request was created by  @slava77 (Slava Krutelyov) for CMSSW_10_6_X.

It involves the following packages:

CommonTools/CandAlgos
PhysicsTools/PatAlgos
PhysicsTools/PatUtils
RecoMET/METAlgorithms


@perrotta, @cmsbuild, @santocch, @slava77 can you please review it and eventually sign? Thanks.
@rappoccio, @gouskos, @hatakeyamak, @emilbols, @peruzzim, @seemasharmafnal, @mmarionncern, @JyothsnaKomaragiri, @makortel, @smoortga, @jdolen, @ferencek, @jdamgov, @nhanvtran, @gkasieczka, @schoef, @andrzejnovak, @clelange, @riga, @ahinzmann, @mverzett, @gpetruc, @mariadalfonso this is something you requested to watch as well.
@silviodonato, @dpiparo you are the release manager for this.

cms-bot commands are listed <a href="http://cms-sw.github.io/cms-bot-cmssw-cmds.html">here</a>