@cmsbuild 2020-05-20
+1
Tested at: 89e72b5dd577528ffb065bf93da1dee2361c770e
https://cmssdt.cern.ch/SDT/jenkins-artifacts/pull-request-integration/PR-0a5455/6449/summary.html
CMSSW: CMSSW_10_6_X_2020-05-19-2300
SCRAM_ARCH: slc7_amd64_gcc700